+++ generic/tkEvent.c
@@ -536,7 +536,7 @@ RefreshKeyboardMappingIfNeeded(
  *----------------------------------------------------------------------
  */
 
-static const int buttonMasks[] = {
+static const unsigned long buttonMasks[] = {
     0, Button1Mask, Button2Mask, Button3Mask, Button4Mask, Button5Mask
 };
